revert change (autoboxing)

+++ copl/com.elster.jupiter.orm/src/com/elster/jupiter/orm/fields/impl/ColumnConversionImpl.java
@@ -47,7 +47,7 @@ public enum ColumnConversionImpl {
 	NUMBER2INTNULLZERO { // 2
 		@Override
 		public Object convertToDb(Object value) {
-			return value == 0 ? null : value;
+			return ((Integer) value) == 0 ? null : value;
 		}
 		
 		@Override
@@ -80,7 +80,7 @@ public enum ColumnConversionImpl {
 	NUMBER2LONGNULLZERO { // 4
 		@Override
 		public Object convertToDb(Object value) {
-			return value == 0 ? null : value;
+			return ((Long) value) == 0 ? null : value;
 		}
 		
 		@Override
